content_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
orjson==3.10.3
elasticsearch[async]==8.13.2
redis==5.0.4
pydantic-settings>=2.0.3
pydantic==2.11.7
python-dotenv==1.1.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-elasticsearch